chore: restructure utils files
