+++ src/pages/ViewProposal.tsx
@@ -16,7 +16,7 @@ import {
 } from '@/src/components/governance/ProposalCard';
 import { HiChevronLeft, HiOutlineClock } from 'react-icons/hi2';
 import { Link } from '@/src/components/ui/Link';
-import { countdownText } from '@/src/lib/date-utils';
+import { countdownText } from '@/src/lib/utils/date';
 import { ProposalResources } from '@/src/components/proposal/ProposalResources';
 import ProposalVotes from '@/src/components/proposal/ProposalVotes';
 import ProposalHistory from '@/src/components/proposal/ProposalHistory';
